Add an alternate Dockerfile for building library images
This is the same as the original Dockerfile, but changes the build tags
to build without rados. The official image will exclude rados support
because the dependencies are awkward to include in the official image.

Add -tags netgo -installsuffix netgo to the Makefile so that the
registry binary won't link to libc unless cgo is involved.

Signed-off-by: Aaron Lehmann <aaron.lehmann@docker.com>